+++ freq.py
@@ -8,40 +8,8 @@ import sys
 from pathlib import Path
 from shutil import rmtree
 
-from config import config, calc_dir
 import obabel_convert
-from run import run_xtb
-
-
-def frequencies(
-    geom_file: Path,
-    charge: int = 0,
-    multiplicity: int = 1,
-    solvation: str | None = None,
-    method: int = 2,
-) -> Path:
-    """Calculate vibrational frequencies and return Gaussian 98 format output file."""
-    unpaired_e = multiplicity - 1
-    command = [
-        "xtb",
-        geom_file,
-        "--hess",
-        "--chrg",
-        str(charge),
-        "--uhf",
-        str(unpaired_e),
-        "--gfn",
-        str(method),
-    ]
-
-    # Add solvation if requested
-    if solvation is not None:
-        command.extend(["--alpb", solvation])
-    # Run xtb from command line
-    calc, out_file, energy = run_xtb(command, geom_file)
-
-    # Return the path of the Gaussian file generated
-    return geom_file.with_name("g98.out")
+from py_xtb import config, calc_dir, calc
 
 
 if __name__ == "__main__":
@@ -80,7 +48,7 @@ if __name__ == "__main__":
             xyz_file.write(str(geom))
 
         # Run calculation; returns path to Gaussian file containing frequencies
-        result_path = frequencies(
+        result_path = calc.frequencies(
             xyz_path,
             charge=avo_input["charge"],
             multiplicity=avo_input["spin"],
