+++ md.py
@@ -8,41 +8,7 @@ import sys
 from pathlib import Path
 from shutil import rmtree, copytree
 
-from config import config, calc_dir
-from run import run_xtb
-
-
-def md(
-    geom_file: Path,
-    input_file: Path,
-    charge: int = 0,
-    multiplicity: int = 1,
-    solvation: str | None = None,
-) -> Path:
-    """Carry out molecular dynamics simulation and return resulting trajectory as multi-geometry xyz-style .trj file."""
-    spin = multiplicity - 1
-    command = [
-        "xtb",
-        geom_file,
-        "--input",
-        input_file,
-        "--omd",
-        "--chrg",
-        str(charge),
-        "--uhf",
-        str(spin),
-    ]
-    # Add solvation if requested
-    if solvation is not None:
-        command.append("--alpb")
-        command.append(solvation)
-    # Run xtb from command line
-    calc, out_file, energy = run_xtb(command, geom_file)
-    # Return path to trajectory file, along with energy
-    return geom_file.with_name("xtb.trj")
-
-
-# def parse_trajectory(trj_file):
+from py_xtb import config, calc_dir, calc
 
 
 if __name__ == "__main__":
@@ -156,7 +122,7 @@ if __name__ == "__main__":
             input_file.write("$end")
 
         # Run calculation using xyz file and input file
-        trj_path = md(
+        trj_path = calc.md(
             xyz_path,
             input_path,
             charge=avo_input["charge"],
